refactor: move TopAppBar to components
